// src/app/shared/enums/validation-fields.ts
export enum ValidationFields {
    // #region contact form
    
    FIRSTNAME = 'firstname',
    LASTNAME = 'lastname',
    EMAIL = 'email',
    PHONE = 'phone',

    // #endregion contact form
    
    // #region task form
    
    TITLE = 'title',
    DESCRIPTION = 'description',
    DUEDATE = 'due date',

    // #endregion task form

    // #region subtask form

    SUBTASK = 'subtask',

    // #endregion subtask form

    // #region sign in form
    // #endregion sign in form

    // #region login form
    // #endregion login form
}